Making lines and stuff

+++ Software/src/C_Dir/device/edk_driver.h
@@ -117,8 +117,8 @@ typedef struct
 
 struct pt
 {
-  uint16_t x;
-  uint16_t y;
+  uint32_t x;
+  uint32_t y;
 };
